Reorganize play actions under sub directory
