+++ server/game/PlayActions/SetupCardAction.js
@@ -1,5 +1,5 @@
-const BaseAbility = require('./baseability.js');
-const Costs = require('./costs.js');
+const BaseAbility = require('../baseability');
+const Costs = require('../costs');
 
 class SetupCardAction extends BaseAbility {
     constructor() {
